+++ src/test/java/io/sui/crypto/SECP256K1DeterministicKeyTest.java
@@ -14,7 +14,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.sui.Account;
+package io.sui.crypto;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
